local_police/requirements.txt
Flask
supabase
pandas
python-dotenv
